feat: Solve issues with latest Ollama version and embedding models
- Limits the character length to 1900 to prevent the models from
  breaking
  - Circumvents #7288 with Ollama
  	- https://github.com/ollama/ollama/issues/7288

Recent versions of Ollama appear to break automatic token truncation for some embedding models, causing certain embedding models to fail unexpectedly. Instead of truncating the input prompt gracefully, Ollama crashes or terminates the connection.

Manual character-length truncation is necessary for the following models:
- `jeffh/intfloat-multilingual-e5-large-instruct:f32`
- `granite-embedding:278m`
- `bge-large:latest`

Without manual limits, Ollama produces warnings for the other models like (this was the case before for all models):
```
  time=2025-04-05T20:06:09.092+02:00 level=WARN source=runner.go:131 msg="truncating input prompt" limit=128 prompt=130 keep=1 new=128
```

If manual limits are not set for the listed embedding models it crashes with:
```
  GGML_ASSERT(i01 >= 0 && i01 < ne01) failed
  [GIN] 2025/04/06 - 09:18:16 | 500 | ... | POST "/api/embed"
  level=ERROR source=server.go:449 msg="llama runner terminated" error="exit status 0xc0000409"
```

Work in progress: diagnosing the issue and considering whether a downgrade of Ollama version is necessary.